+++ hutool-core/src/main/java/cn/hutool/core/thread/GlobalThreadPool.java
@@ -7,7 +7,8 @@ import java.util.concurrent.Future;
 import cn.hutool.core.exceptions.UtilException;
 
 /**
- * 全局公共线程池
+ * 全局公共线程池<br>
+ * 此线程池是一个无限线程池，既加入的线程不等待任何线程，直接执行
  * 
  * @author Looly
  *