+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * 【cron】          CronPatternUtil增加nextDateAfter方法（issue#IVYNL@Github）
 * 【core】          增加RandomUtil.randomDate方法（issue#IW49T@Github）
 * 【db】             Table增加comment字段，调整元信息逻辑（issue#IW49S@Gitee）
+* 【core】          增加ConcurrencyTester（pr#41@Gitee）
 
 ### Bug修复
 * 【core】          修复Enjoy模板创建多个引擎报错问题（issue#344@Github）